requirements.txt
streamlit
google-generativeai
pymongo
pypdf2
python-dotenv
pyvis
streamlit-pyvis
